requirements.txt
asgiref==3.8.1
boto3==1.36.23
botocore==1.36.23
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
dj-database-url==2.3.0
Django==5.1.6
django-cognito-jwt==0.0.4
django-decouple==2.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
jmespath==1.0.1
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
